refactor: moving around filters and hiding reply arrows
