+++ lib/views/home/chat/widgets/message-list.dart
@@ -7,6 +7,7 @@ import 'package:syphon/global/colours.dart';
 import 'package:syphon/global/print.dart';
 import 'package:syphon/store/events/actions.dart';
 import 'package:syphon/store/events/messages/model.dart';
+import 'package:syphon/store/events/messages/selectors.dart';
 import 'package:syphon/store/events/reactions/actions.dart';
 import 'package:syphon/store/events/selectors.dart';
 import 'package:syphon/store/index.dart';
